// HTML5/dev/simplivity-citrixplugin-ui/src/app/model/vmdata.model.ts
/* Copyright (c) 2018 VMware, Inc. All rights reserved. */

export class VMData {
   vmName: string;
   vmUsername: string;
   vmPassword: string;
   vmCluster: string;
   vmHost: string;
   isStatic: boolean;
   ipAddr: string;
   subnet: string;
   gateway: string;
   dnsServerAddress: string;
}